+++ src/lib/jobs.rs
@@ -14,18 +14,6 @@ use synchronoise::SignalEvent;
 use tokio::sync::{mpsc as tokio_mpsc, Mutex};
 use tonic::Status;
 
-/*
-Tonic requires a bounded tokio mpsc stream to return to the streaming client.
-You need to ensure a large enough buffer size so the sender does not return
-an error. A size of 4096 should not be necessary but is used here as a safe guard.
-
-TODO:
-Run multiple client streaming tests to determine the maximum number of items
-that appeared in the buffer at once. Set the STREAM_BUFFER_SIZE to be
-1.5 times larger than this.
-*/
-const STREAM_BUFFER_SIZE: usize = 4096;
-
 /// A user job containing information about the
 /// underlying process.
 pub struct Job {
