+++ src/lib/lib.rs
@@ -5,4 +5,5 @@ mod errors;
 pub mod jobs;
 mod processing;
 pub mod server;
+mod server_types;
 mod utils;
